client/pyproject.toml
[project]
name = "joinly-client"
version = "0.1.0"
description = "Client for joinly: Make your meetings accessible to AI Agents"
readme = "README.md"
authors = [
    { name = "dbrockmann", email = "dominik@joinly.ai" }
]
requires-python = ">=3.12"
dependencies = [
    "click>=8.1.8",
    "fastmcp>=2.10.1",
    "langchain>=0.3.25",
    "langchain-anthropic>=0.3.15",
    "langchain-mcp-adapters>=0.0.3",
    "langchain-ollama>=0.3.3",
    "langchain-openai>=0.3.16",
    "langgraph>=0.4.1",
    "python-dotenv>=1.1.0",
    "rich>=14.0.0",
]

[project.scripts]
joinly-client = "joinly_client.main:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["joinly-client"]

[tool.hatch.build.targets.wheel.force-include]
"../shared" = "shared"

[project.urls]
"Repository" = "https://github.com/joinly-ai/joinly"
